+++ selfservice/flow/registration/hook.go
@@ -7,7 +7,6 @@ import (
 	"context"
 	"fmt"
 	"net/http"
-	"net/url"
 	"time"
 
 	"go.opentelemetry.io/otel/trace"
@@ -249,22 +248,15 @@ func (e *HookExecutor) PostRegistrationHook(w http.ResponseWriter, r *http.Reque
 
 	finalReturnTo := returnTo.String()
 	if a.OAuth2LoginChallenge != "" {
-		callbackURL, err := e.d.Hydra().AcceptLoginRequest(r.Context(), string(a.OAuth2LoginChallenge), i.ID.String(), s.AMR)
-		if err != nil {
-			return err
-		}
 		if a.ReturnToVerification != "" {
 			// Special case: If Kratos is used as a login UI *and* we want to show the verification UI,
 			// redirect to the verification URL first and then return to Hydra.
-			verificationURL, err := url.Parse(a.ReturnToVerification)
+			finalReturnTo = a.ReturnToVerification
+		} else {
+			callbackURL, err := e.d.Hydra().AcceptLoginRequest(r.Context(), string(a.OAuth2LoginChallenge), i.ID.String(), s.AMR)
 			if err != nil {
 				return err
 			}
-			q := verificationURL.Query()
-			q.Set("return_to", callbackURL)
-			verificationURL.RawQuery = q.Encode()
-			finalReturnTo = verificationURL.String()
-		} else {
 			finalReturnTo = callbackURL
 		}
 	} else if a.ReturnToVerification != "" {